Print the path to the generated `.dart` file on success
Provide better feedback about what `resource_importer` did by
printing the path to the generated `.dart` file.

This was slightly non-trivial because I did not want to generate
additional spew in tests.

Also try to ensure that error messages include the application name
to make failures easier to diagnose if called as part of larger
scripts.