Write a checksum of the configuration into the generated file
Write a checksum of the configuration into the generated file so that
we can eventually can avoid unnecessary code regeneration.

I debated whether to compute the checksum with CRC32, MD5, SHA-1, or
SHA-256.  The checksum does not need to be cryptographically secure.
I went with SHA-1 because it's common, and even though CRC32 and MD5
are much faster, the difference is unlikely to matter in practice for
what typically should be small inputs.